Rewrite main part in C
