+++ raycaster_data.c
@@ -1,0 +1,307 @@
+#include <stdint.h>
+#include "raycaster.h"
+
+const uint8_t LOOKUP_TBL g_map[] = {
+    0b00000000, 0b10000000, 0b00000000, 0b00000000, 0b01111010, 0b10111111,
+    0b11111111, 0b00000000, 0b00111000, 0b10100000, 0b00001000, 0b01001100,
+    0b01000001, 0b00000100, 0b00100100, 0b00001100, 0b00000000, 0b10001010,
+    0b00000010, 0b01011100, 0b10000001, 0b11000100, 0b00000110, 0b00001100,
+    0b00000000, 0b11010010, 0b00010000, 0b00001100, 0b01001100, 0b10010000,
+    0b00000000, 0b00111100,
+
+    0b00100011, 0b10100100, 0b00000100, 0b00001100, 0b00000000, 0b11000001,
+    0b10001000, 0b00001100, 0b00110000, 0b10011100, 0b00111000, 0b01111100,
+    0b00000011, 0b00000000, 0b00000000, 0b00001100, 0b00011111, 0b11111110,
+    0b11111111, 0b11111100, 0b00000000, 0b00000000, 0b00000000, 0b00000000,
+    0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000,
+    0b00000000, 0b00000000,
+
+    0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000,
+    0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000,
+    0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000,
+    0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000,
+    0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000,
+    0b00000000, 0b00000000,
+
+    0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000,
+    0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000,
+    0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000,
+    0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000,
+    0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000, 0b00000000,
+    0b00000000, 0b00000000};
+
+const uint8_t LOOKUP_TBL g_texture8[4096] = {
+    44,  44,  44,  44,  44,  44,  44,  44,  44,  44,  44,  44,  44,  44,  44,
+    44,  44,  44,  44,  44,  44,  44,  44,  44,  44,  44,  44,  44,  44,  44,
+    44,  44,  44,  44,  44,  44,  44,  44,  44,  44,  44,  44,  44,  44,  44,
+    44,  44,  44,  44,  44,  44,  44,  44,  44,  44,  44,  44,  44,  44,  44,
+    44,  44,  44,  44,  84,  84,  84,  84,  84,  84,  84,  84,  84,  84,  72,
+    220, 220, 220, 220, 220, 220, 220, 220, 220, 220, 220, 220, 220, 220, 220,
+    220, 32,  56,  140, 168, 192, 208, 208, 208, 208, 208, 208, 192, 192, 168,
+    140, 84,  32,  72,  72,  72,  72,  72,  72,  72,  72,  72,  72,  84,  84,
+    84,  84,  84,  84,  84,  84,  84,  84,  84,  84,  84,  84,  84,  84,  84,
+    84,  84,  84,  72,  220, 208, 208, 208, 192, 208, 208, 208, 208, 208, 208,
+    192, 208, 208, 208, 140, 32,  56,  140, 192, 180, 140, 168, 140, 140, 140,
+    140, 140, 140, 168, 84,  56,  32,  56,  112, 140, 140, 140, 140, 140, 140,
+    124, 112, 100, 72,  72,  72,  72,  72,  72,  72,  72,  72,  140, 140, 140,
+    140, 140, 140, 140, 140, 32,  84,  72,  220, 208, 192, 180, 208, 180, 208,
+    192, 180, 208, 192, 192, 180, 208, 140, 100, 32,  56,  168, 220, 168, 140,
+    168, 140, 168, 140, 168, 140, 168, 140, 84,  56,  32,  56,  140, 220, 220,
+    220, 220, 220, 220, 192, 180, 168, 140, 140, 140, 140, 140, 140, 140, 140,
+    140, 208, 220, 208, 208, 180, 168, 140, 140, 32,  84,  72,  220, 208, 168,
+    168, 168, 168, 140, 168, 168, 168, 168, 152, 168, 168, 124, 84,  32,  56,
+    140, 192, 140, 140, 140, 168, 168, 168, 168, 140, 140, 112, 84,  56,  32,
+    56,  140, 220, 220, 192, 192, 192, 208, 220, 220, 220, 220, 220, 208, 208,
+    208, 208, 208, 208, 208, 168, 168, 168, 168, 168, 168, 124, 112, 32,  84,
+    72,  236, 192, 152, 168, 168, 168, 168, 168, 168, 152, 168, 140, 168, 168,
+    112, 84,  32,  56,  72,  168, 140, 168, 168, 140, 140, 168, 140, 168, 112,
+    84,  84,  56,  32,  56,  140, 220, 180, 168, 140, 168, 140, 168, 168, 168,
+    140, 140, 168, 140, 168, 140, 168, 140, 152, 180, 152, 168, 152, 168, 140,
+    124, 112, 32,  72,  72,  236, 208, 168, 140, 168, 152, 168, 140, 152, 152,
+    152, 140, 152, 168, 140, 84,  32,  56,  72,  124, 140, 140, 168, 140, 168,
+    168, 140, 168, 112, 56,  56,  56,  32,  56,  140, 220, 180, 140, 168, 140,
+    152, 140, 140, 140, 168, 140, 140, 152, 152, 168, 152, 140, 168, 180, 180,
+    124, 180, 140, 152, 100, 84,  32,  56,  72,  236, 208, 140, 152, 140, 168,
+    168, 168, 140, 124, 140, 152, 168, 140, 100, 84,  32,  56,  72,  72,  140,
+    168, 168, 168, 140, 140, 140, 168, 112, 56,  56,  56,  44,  56,  140, 220,
+    180, 168, 140, 152, 124, 168, 152, 124, 140, 168, 140, 168, 168, 140, 124,
+    168, 152, 152, 168, 168, 152, 152, 112, 72,  72,  32,  44,  72,  192, 192,
+    168, 152, 168, 140, 168, 140, 152, 140, 152, 140, 140, 168, 112, 84,  32,
+    32,  72,  84,  72,  124, 56,  56,  56,  56,  56,  56,  56,  56,  56,  32,
+    44,  56,  124, 220, 220, 140, 168, 140, 124, 152, 124, 152, 140, 168, 140,
+    152, 124, 168, 168, 124, 140, 140, 140, 140, 140, 112, 84,  72,  72,  32,
+    44,  72,  168, 236, 168, 168, 168, 140, 140, 152, 140, 140, 140, 140, 168,
+    140, 84,  124, 44,  32,  56,  72,  56,  56,  56,  56,  56,  56,  56,  56,
+    56,  32,  32,  32,  72,  72,  72,  192, 208, 140, 140, 140, 152, 152, 152,
+    152, 140, 140, 168, 140, 168, 168, 140, 168, 140, 112, 112, 112, 112, 84,
+    72,  56,  56,  32,  44,  72,  140, 220, 140, 168, 140, 152, 124, 140, 140,
+    168, 140, 168, 140, 168, 112, 84,  44,  32,  56,  72,  56,  44,  44,  32,
+    32,  32,  32,  32,  44,  44,  44,  44,  72,  84,  72,  192, 180, 140, 168,
+    140, 168, 140, 140, 140, 112, 112, 100, 100, 100, 100, 100, 100, 100, 84,
+    84,  72,  72,  72,  32,  32,  32,  32,  44,  72,  112, 220, 168, 140, 152,
+    124, 124, 124, 168, 152, 140, 140, 124, 140, 112, 100, 56,  32,  56,  72,
+    72,  72,  72,  112, 124, 140, 168, 168, 168, 168, 152, 112, 56,  84,  72,
+    180, 140, 112, 112, 100, 100, 100, 84,  84,  84,  84,  72,  72,  72,  72,
+    72,  72,  72,  32,  32,  32,  32,  32,  44,  44,  44,  44,  56,  84,  72,
+    220, 208, 140, 140, 140, 152, 152, 140, 140, 168, 124, 140, 168, 140, 84,
+    56,  32,  56,  72,  112, 168, 168, 220, 220, 236, 220, 208, 208, 152, 112,
+    84,  32,  84,  72,  140, 112, 100, 100, 84,  72,  72,  72,  56,  56,  56,
+    56,  56,  44,  44,  44,  44,  44,  44,  44,  44,  44,  44,  44,  44,  44,
+    44,  56,  84,  72,  168, 208, 140, 168, 140, 124, 140, 168, 140, 140, 152,
+    140, 140, 112, 84,  56,  32,  44,  56,  168, 192, 220, 208, 180, 140, 168,
+    140, 168, 140, 112, 56,  32,  56,  56,  56,  56,  56,  56,  56,  56,  56,
+    56,  56,  44,  44,  32,  32,  32,  32,  32,  32,  32,  56,  56,  56,  56,
+    56,  56,  56,  56,  56,  56,  84,  72,  140, 208, 168, 140, 168, 152, 140,
+    124, 140, 124, 140, 168, 140, 84,  84,  56,  32,  44,  72,  168, 220, 192,
+    168, 140, 140, 140, 140, 140, 140, 84,  56,  32,  56,  44,  44,  44,  44,
+    44,  44,  44,  44,  44,  44,  56,  56,  56,  56,  56,  56,  56,  56,  56,
+    84,  72,  56,  56,  56,  56,  56,  56,  56,  56,  84,  72,  112, 192, 140,
+    168, 152, 140, 168, 152, 168, 140, 168, 140, 168, 84,  112, 72,  32,  44,
+    72,  168, 236, 168, 140, 140, 168, 140, 168, 140, 168, 84,  56,  32,  56,
+    72,  84,  84,  84,  84,  32,  72,  72,  72,  72,  72,  72,  72,  72,  72,
+    72,  72,  72,  72,  152, 152, 152, 152, 152, 152, 152, 152, 152, 44,  32,
+    84,  72,  192, 192, 152, 140, 140, 152, 152, 124, 140, 140, 140, 140, 112,
+    84,  84,  32,  32,  72,  168, 236, 140, 140, 168, 168, 140, 168, 140, 140,
+    84,  56,  32,  56,  112, 168, 220, 220, 84,  72,  84,  124, 152, 152, 152,
+    152, 152, 152, 152, 152, 152, 152, 152, 208, 208, 208, 208, 208, 208, 180,
+    180, 112, 56,  32,  84,  72,  140, 192, 140, 168, 140, 124, 140, 140, 140,
+    168, 140, 112, 140, 112, 72,  32,  32,  72,  168, 208, 168, 168, 140, 140,
+    168, 140, 168, 140, 84,  56,  32,  56,  168, 220, 192, 168, 32,  72,  124,
+    192, 220, 220, 220, 220, 220, 220, 220, 220, 220, 220, 220, 168, 152, 140,
+    168, 140, 168, 168, 112, 112, 56,  32,  84,  72,  112, 152, 168, 140, 168,
+    168, 140, 168, 112, 140, 112, 84,  112, 84,  72,  32,  32,  72,  168, 208,
+    168, 140, 112, 112, 112, 112, 84,  84,  84,  56,  32,  56,  220, 192, 168,
+    140, 32,  72,  152, 220, 208, 180, 180, 180, 180, 180, 180, 180, 180, 180,
+    180, 168, 140, 168, 140, 168, 140, 140, 112, 112, 56,  32,  72,  84,  72,
+    152, 152, 112, 112, 112, 112, 84,  84,  112, 84,  112, 84,  72,  44,  32,
+    32,  44,  168, 140, 112, 140, 112, 56,  56,  56,  56,  56,  56,  56,  32,
+    100, 220, 180, 140, 140, 32,  72,  152, 220, 180, 168, 152, 168, 152, 168,
+    140, 152, 140, 152, 152, 168, 152, 140, 168, 140, 168, 168, 112, 84,  56,
+    32,  72,  84,  72,  100, 84,  84,  84,  84,  72,  72,  72,  56,  44,  32,
+    32,  32,  32,  32,  32,  44,  140, 112, 56,  56,  56,  32,  32,  32,  32,
+    32,  32,  32,  32,  100, 220, 168, 168, 140, 32,  72,  152, 220, 168, 140,
+    152, 140, 152, 140, 152, 124, 124, 140, 140, 168, 140, 168, 140, 168, 140,
+    140, 112, 84,  56,  32,  72,  84,  72,  44,  44,  44,  44,  32,  32,  32,
+    32,  32,  32,  32,  32,  32,  32,  32,  32,  44,  56,  84,  56,  32,  32,
+    32,  32,  44,  44,  44,  44,  44,  44,  140, 220, 168, 168, 140, 32,  72,
+    140, 192, 140, 168, 152, 168, 152, 168, 140, 152, 140, 152, 152, 140, 168,
+    124, 124, 140, 140, 168, 112, 84,  56,  32,  56,  84,  84,  72,  56,  32,
+    32,  32,  32,  32,  32,  32,  32,  72,  72,  72,  72,  72,  72,  72,  140,
+    140, 152, 168, 168, 152, 140, 140, 124, 84,  44,  56,  100, 192, 220, 168,
+    140, 124, 32,  72,  140, 192, 168, 140, 152, 140, 152, 140, 152, 124, 124,
+    140, 140, 140, 140, 124, 124, 140, 168, 140, 100, 72,  56,  32,  56,  140,
+    220, 220, 220, 220, 192, 180, 168, 140, 100, 56,  72,  100, 124, 168, 168,
+    192, 192, 208, 220, 220, 220, 220, 220, 220, 220, 208, 152, 100, 32,  56,
+    124, 220, 192, 168, 168, 112, 32,  72,  124, 192, 168, 152, 140, 168, 140,
+    168, 140, 124, 124, 140, 152, 140, 140, 168, 168, 140, 140, 140, 100, 72,
+    56,  32,  56,  220, 192, 168, 152, 152, 152, 168, 152, 152, 56,  84,  72,
+    124, 208, 236, 236, 220, 192, 180, 192, 192, 192, 192, 192, 192, 192, 192,
+    112, 84,  32,  56,  140, 220, 180, 152, 140, 112, 32,  72,  112, 192, 140,
+    168, 152, 140, 152, 140, 140, 168, 140, 168, 140, 152, 168, 140, 140, 168,
+    140, 112, 84,  56,  56,  32,  56,  220, 168, 140, 152, 168, 168, 152, 152,
+    168, 32,  84,  72,  152, 236, 192, 140, 168, 140, 168, 152, 140, 168, 152,
+    168, 152, 140, 140, 100, 72,  32,  56,  140, 236, 168, 140, 168, 84,  32,
+    72,  72,  140, 112, 100, 84,  84,  84,  100, 100, 112, 124, 140, 140, 100,
+    100, 100, 100, 100, 84,  72,  56,  56,  56,  32,  56,  220, 168, 168, 168,
+    140, 168, 140, 168, 100, 32,  72,  72,  152, 236, 168, 168, 168, 140, 124,
+    168, 168, 140, 140, 140, 168, 140, 168, 84,  72,  32,  56,  140, 236, 168,
+    168, 152, 84,  32,  84,  32,  56,  56,  72,  72,  72,  72,  72,  72,  72,
+    84,  84,  84,  84,  72,  72,  56,  56,  56,  32,  32,  32,  32,  44,  56,
+    220, 140, 168, 140, 168, 180, 140, 124, 100, 32,  72,  72,  152, 236, 140,
+    168, 140, 152, 168, 140, 140, 124, 140, 168, 152, 140, 140, 84,  72,  32,
+    56,  140, 236, 140, 168, 140, 84,  32,  84,  56,  32,  32,  32,  32,  32,
+    32,  32,  32,  32,  56,  56,  72,  32,  32,  32,  32,  32,  32,  44,  44,
+    44,  72,  72,  72,  220, 168, 152, 140, 168, 140, 168, 124, 100, 32,  56,
+    72,  152, 236, 152, 140, 168, 140, 140, 124, 140, 124, 140, 140, 140, 168,
+    140, 84,  72,  32,  56,  140, 236, 168, 168, 140, 84,  32,  84,  56,  56,
+    56,  56,  56,  56,  56,  44,  44,  44,  32,  32,  32,  72,  100, 100, 112,
+    140, 152, 140, 140, 140, 140, 72,  32,  220, 168, 124, 152, 140, 168, 140,
+    124, 100, 56,  32,  72,  152, 236, 168, 140, 140, 140, 124, 140, 124, 140,
+    168, 124, 140, 140, 168, 84,  72,  32,  56,  140, 236, 152, 168, 168, 84,
+    32,  84,  72,  72,  72,  84,  84,  84,  84,  84,  84,  84,  56,  56,  56,
+    100, 140, 180, 220, 220, 220, 208, 208, 208, 168, 112, 32,  220, 168, 168,
+    140, 124, 140, 168, 140, 100, 56,  32,  72,  140, 220, 192, 152, 140, 168,
+    140, 168, 124, 124, 140, 124, 140, 168, 140, 84,  56,  32,  56,  140, 220,
+    140, 140, 168, 84,  32,  72,  168, 220, 220, 180, 140, 100, 84,  72,  72,
+    72,  72,  72,  84,  100, 180, 208, 208, 192, 180, 168, 168, 152, 168, 72,
+    32,  236, 152, 152, 168, 124, 140, 168, 152, 112, 72,  32,  72,  124, 208,
+    192, 168, 140, 140, 124, 140, 124, 124, 168, 140, 168, 140, 124, 84,  56,
+    44,  56,  140, 220, 168, 152, 140, 84,  32,  72,  220, 140, 140, 180, 220,
+    220, 220, 220, 220, 220, 152, 56,  84,  100, 220, 180, 168, 168, 140, 168,
+    152, 168, 140, 112, 32,  236, 152, 168, 152, 168, 168, 140, 168, 112, 72,
+    32,  56,  84,  192, 236, 168, 168, 152, 140, 168, 140, 168, 124, 124, 140,
+    168, 112, 84,  56,  44,  56,  140, 220, 140, 168, 140, 84,  32,  72,  220,
+    140, 140, 168, 168, 140, 140, 168, 140, 140, 124, 56,  84,  100, 220, 180,
+    152, 140, 168, 140, 168, 140, 168, 112, 32,  236, 152, 140, 152, 168, 140,
+    168, 140, 100, 84,  32,  44,  84,  168, 236, 168, 168, 168, 140, 140, 124,
+    140, 124, 124, 168, 140, 112, 72,  44,  44,  72,  140, 208, 168, 168, 140,
+    84,  32,  72,  220, 168, 168, 140, 140, 168, 168, 140, 168, 140, 100, 56,
+    72,  100, 220, 168, 152, 168, 140, 168, 168, 124, 168, 84,  32,  220, 168,
+    168, 168, 152, 140, 152, 168, 100, 84,  32,  44,  84,  140, 236, 192, 168,
+    140, 168, 140, 140, 152, 140, 168, 140, 168, 100, 72,  32,  44,  72,  140,
+    208, 168, 140, 168, 84,  32,  72,  208, 168, 140, 168, 168, 140, 140, 168,
+    140, 168, 100, 56,  72,  100, 220, 180, 168, 140, 140, 168, 152, 168, 140,
+    84,  32,  220, 192, 124, 152, 124, 168, 168, 152, 100, 84,  32,  32,  72,
+    112, 208, 192, 168, 140, 124, 152, 140, 140, 152, 140, 152, 140, 100, 56,
+    32,  44,  72,  140, 208, 140, 168, 84,  84,  32,  72,  192, 168, 168, 168,
+    140, 168, 140, 140, 168, 168, 84,  56,  72,  100, 220, 180, 140, 168, 168,
+    140, 168, 140, 168, 84,  32,  192, 220, 140, 152, 168, 140, 152, 140, 100,
+    84,  32,  32,  72,  84,  180, 220, 168, 140, 168, 140, 168, 140, 168, 140,
+    140, 100, 84,  56,  32,  56,  72,  124, 168, 124, 84,  56,  56,  56,  72,
+    168, 168, 140, 168, 168, 140, 168, 140, 140, 168, 84,  56,  56,  100, 220,
+    180, 168, 168, 140, 168, 152, 168, 140, 84,  32,  180, 220, 124, 168, 140,
+    140, 140, 168, 112, 84,  32,  32,  56,  84,  140, 192, 168, 140, 124, 100,
+    100, 84,  72,  72,  56,  56,  56,  56,  32,  56,  72,  84,  56,  56,  56,
+    56,  72,  72,  32,  112, 84,  72,  84,  84,  84,  84,  84,  84,  84,  84,
+    56,  56,  100, 220, 180, 140, 168, 168, 140, 168, 124, 168, 84,  32,  152,
+    220, 152, 124, 168, 168, 140, 140, 112, 84,  32,  32,  56,  84,  72,  140,
+    140, 100, 84,  72,  56,  32,  32,  32,  32,  32,  32,  32,  32,  56,  72,
+    84,  72,  72,  84,  84,  84,  84,  84,  32,  32,  32,  32,  32,  32,  32,
+    32,  32,  32,  32,  32,  56,  100, 220, 180, 140, 140, 152, 168, 140, 168,
+    140, 84,  32,  124, 220, 168, 152, 168, 140, 140, 168, 100, 84,  32,  32,
+    44,  72,  72,  56,  44,  44,  44,  32,  32,  44,  44,  44,  44,  44,  44,
+    44,  44,  56,  72,  84,  84,  72,  72,  72,  72,  72,  72,  72,  72,  72,
+    56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  100, 220, 180, 168, 168,
+    140, 152, 168, 152, 168, 84,  32,  84,  236, 140, 168, 140, 140, 168, 140,
+    100, 84,  32,  32,  44,  72,  72,  72,  56,  56,  56,  56,  56,  72,  72,
+    72,  72,  72,  72,  72,  72,  72,  72,  84,  84,  124, 124, 124, 124, 124,
+    124, 124, 124, 124, 124, 124, 124, 124, 124, 124, 124, 124, 56,  84,  100,
+    220, 180, 140, 140, 168, 140, 168, 140, 112, 84,  32,  72,  236, 168, 140,
+    140, 124, 140, 168, 112, 84,  44,  32,  44,  56,  100, 124, 124, 124, 124,
+    124, 124, 124, 124, 124, 124, 124, 124, 124, 140, 140, 152, 152, 168, 180,
+    180, 192, 192, 192, 192, 192, 192, 180, 192, 192, 192, 192, 192, 192, 192,
+    124, 56,  84,  100, 208, 180, 168, 168, 152, 168, 140, 168, 84,  32,  44,
+    72,  236, 168, 140, 168, 140, 140, 140, 100, 84,  44,  32,  32,  56,  124,
+    192, 192, 192, 192, 192, 192, 192, 192, 192, 192, 192, 192, 192, 192, 192,
+    208, 220, 220, 236, 236, 236, 236, 236, 236, 236, 236, 236, 236, 236, 220,
+    220, 208, 208, 152, 100, 44,  72,  100, 208, 180, 152, 168, 140, 168, 168,
+    140, 84,  32,  44,  72,  236, 140, 168, 124, 152, 140, 168, 100, 84,  56,
+    32,  32,  32,  124, 192, 208, 236, 236, 236, 236, 236, 236, 220, 236, 236,
+    220, 236, 236, 236, 236, 236, 236, 220, 220, 208, 192, 180, 192, 192, 192,
+    168, 192, 192, 180, 192, 192, 180, 124, 84,  44,  72,  100, 192, 180, 168,
+    140, 168, 140, 140, 140, 84,  32,  56,  72,  220, 168, 168, 140, 168, 140,
+    140, 100, 84,  56,  32,  32,  32,  124, 192, 208, 180, 192, 192, 192, 192,
+    192, 168, 192, 192, 168, 192, 192, 192, 180, 180, 168, 152, 152, 140, 140,
+    168, 140, 168, 140, 140, 140, 168, 140, 168, 140, 168, 112, 84,  44,  56,
+    72,  192, 140, 140, 168, 140, 112, 112, 112, 84,  32,  56,  72,  208, 140,
+    140, 168, 140, 168, 140, 112, 84,  56,  32,  32,  32,  100, 192, 220, 180,
+    140, 140, 140, 140, 140, 140, 168, 140, 168, 168, 152, 140, 152, 168, 168,
+    168, 152, 124, 152, 140, 168, 168, 152, 140, 168, 140, 168, 168, 168, 140,
+    112, 84,  44,  56,  72,  180, 168, 168, 168, 140, 84,  84,  84,  32,  44,
+    56,  72,  180, 168, 140, 140, 168, 140, 140, 112, 84,  56,  32,  32,  32,
+    72,  192, 220, 192, 140, 152, 140, 168, 140, 168, 140, 152, 168, 140, 140,
+    152, 168, 168, 140, 152, 140, 152, 140, 140, 124, 140, 152, 152, 140, 168,
+    140, 168, 140, 168, 112, 84,  44,  56,  72,  124, 124, 124, 84,  84,  84,
+    84,  84,  32,  44,  56,  72,  152, 152, 168, 140, 140, 140, 112, 100, 84,
+    56,  32,  32,  32,  44,  180, 220, 208, 168, 168, 140, 140, 140, 140, 124,
+    140, 140, 168, 140, 152, 140, 168, 152, 124, 140, 140, 168, 140, 140, 168,
+    140, 124, 124, 140, 124, 152, 140, 152, 112, 84,  32,  56,  72,  56,  32,
+    32,  32,  84,  84,  84,  84,  32,  56,  72,  72,  124, 100, 100, 100, 100,
+    100, 100, 84,  84,  56,  32,  32,  32,  44,  168, 220, 208, 140, 140, 168,
+    140, 168, 140, 140, 168, 140, 124, 124, 152, 152, 140, 168, 140, 168, 152,
+    140, 168, 140, 140, 124, 140, 140, 140, 168, 140, 124, 124, 112, 84,  32,
+    56,  84,  84,  84,  72,  56,  32,  32,  44,  32,  44,  56,  72,  56,  56,
+    56,  56,  56,  56,  56,  56,  56,  56,  56,  32,  32,  32,  44,  140, 208,
+    180, 180, 140, 140, 168, 140, 168, 124, 140, 140, 152, 124, 152, 140, 152,
+    124, 168, 140, 140, 124, 140, 140, 140, 140, 168, 140, 124, 140, 152, 124,
+    124, 112, 84,  32,  56,  72,  72,  72,  72,  72,  72,  72,  44,  44,  44,
+    56,  72,  56,  56,  32,  32,  32,  32,  32,  32,  32,  32,  32,  32,  32,
+    32,  44,  140, 152, 112, 112, 112, 112, 112, 112, 112, 168, 140, 168, 168,
+    140, 140, 124, 140, 140, 140, 124, 140, 140, 168, 140, 168, 140, 140, 168,
+    140, 140, 140, 140, 152, 112, 84,  32,  56,  72,  124, 168, 168, 168, 168,
+    152, 140, 100, 72,  72,  84,  56,  56,  44,  44,  44,  44,  44,  44,  44,
+    44,  44,  44,  44,  44,  56,  84,  72,  56,  56,  44,  44,  44,  44,  44,
+    84,  112, 112, 124, 140, 124, 124, 124, 124, 124, 124, 124, 124, 124, 124,
+    124, 124, 124, 124, 124, 124, 124, 124, 124, 100, 84,  32,  56,  72,  168,
+    192, 220, 220, 220, 220, 220, 192, 168, 168, 140, 112, 100, 84,  72,  56,
+    56,  56,  56,  56,  56,  56,  56,  56,  56,  72,  84,  72,  72,  56,  56,
+    56,  56,  56,  56,  44,  72,  84,  84,  84,  84,  84,  84,  84,  84,  84,
+    84,  84,  84,  84,  84,  84,  84,  84,  84,  84,  84,  84,  84,  84,  32,
+    44,  56,  72,  168, 220, 168, 168, 140, 192, 208, 220, 220, 220, 192, 168,
+    168, 168, 168, 168, 168, 168, 168, 140, 140, 140, 124, 124, 112, 100, 100,
+    84,  84,  72,  72,  72,  72,  72,  56,  56,  44,  44,  44,  44,  44,  44,
+    44,  44,  44,  44,  44,  44,  44,  32,  32,  32,  32,  32,  32,  32,  32,
+    32,  32,  32,  44,  56,  72,  72,  168, 220, 168, 168, 168, 180, 180, 140,
+    168, 192, 220, 220, 220, 220, 220, 220, 220, 220, 220, 220, 220, 220, 208,
+    208, 192, 192, 168, 168, 168, 168, 168, 140, 72,  84,  72,  56,  56,  56,
+    56,  56,  56,  56,  56,  56,  44,  44,  44,  44,  44,  44,  44,  44,  44,
+    44,  44,  44,  44,  44,  44,  56,  56,  72,  72,  72,  168, 220, 168, 140,
+    168, 124, 124, 168, 152, 140, 124, 168, 140, 168, 140, 168, 140, 168, 168,
+    140, 168, 140, 168, 168, 180, 192, 220, 220, 220, 208, 192, 124, 56,  72,
+    84,  72,  72,  72,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,
+    56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  72,  72,  72,  84,  72,
+    168, 208, 168, 168, 168, 152, 168, 140, 152, 140, 152, 168, 140, 140, 168,
+    124, 124, 140, 168, 124, 140, 140, 140, 168, 140, 168, 140, 168, 168, 168,
+    124, 100, 32,  56,  84,  72,  168, 220, 220, 168, 124, 72,  72,  72,  84,
+    84,  84,  72,  72,  72,  72,  72,  72,  72,  72,  72,  72,  72,  72,  84,
+    84,  84,  84,  72,  168, 192, 168, 140, 168, 124, 124, 152, 168, 140, 168,
+    168, 152, 140, 124, 168, 124, 140, 124, 152, 140, 168, 140, 168, 140, 168,
+    140, 168, 168, 168, 100, 84,  32,  56,  72,  72,  220, 168, 168, 220, 220,
+    220, 220, 180, 84,  84,  72,  180, 220, 220, 220, 220, 220, 220, 220, 208,
+    192, 180, 152, 124, 72,  72,  84,  72,  168, 140, 100, 100, 100, 100, 100,
+    100, 100, 100, 100, 100, 100, 100, 100, 100, 100, 100, 100, 100, 100, 100,
+    100, 100, 100, 100, 100, 100, 100, 100, 84,  84,  56,  32,  56,  72,  220,
+    180, 180, 180, 180, 180, 180, 180, 100, 84,  124, 220, 168, 168, 168, 140,
+    168, 140, 168, 152, 168, 140, 168, 220, 152, 32,  84,  72,  140, 100, 56,
+    56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,
+    56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,
+    32,  44,  72,  180, 168, 180, 180, 152, 100, 152, 140, 100, 72,  180, 168,
+    140, 168, 140, 168, 168, 168, 140, 168, 140, 168, 140, 168, 84,  32,  84,
+    72,  56,  32,  32,  32,  32,  32,  32,  32,  32,  32,  32,  32,  32,  32,
+    32,  32,  32,  32,  32,  32,  32,  32,  32,  32,  32,  32,  32,  32,  32,
+    32,  32,  32,  32,  32,  44,  72,  56,  72,  56,  84,  100, 100, 100, 100,
+    100, 72,  220, 168, 152, 152, 168, 168, 152, 152, 168, 152, 168, 168, 140,
+    140, 84,  32,  84,  84,  72,  56,  56,  56,  44,  44,  44,  44,  44,  44,
+    44,  44,  44,  44,  44,  44,  44,  44,  44,  44,  44,  44,  44,  44,  44,
+    44,  44,  44,  44,  44,  44,  44,  44,  44,  44,  56,  84,  84,  56,  56,
+    32,  32,  32,  32,  32,  56,  84,  84,  84,  84,  84,  84,  84,  84,  84,
+    84,  84,  84,  84,  84,  84,  32,  84,  84,  84,  84,  84,  72,  72,  72,
+    72,  72,  72,  72,  72,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,
+    56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,
+    84,  84,  84,  84,  56,  56,  56,  56,  56,  56,  84,  84,  32,  32,  32,
+    32,  32,  32,  32,  32,  32,  32,  32,  32,  32,  32,  84,  56,  56,  56,
+    56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,
+    56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,
+    56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,
+    56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,  56,
+    56};
